chore(backend): renaming backend -> archaide-backend
