+++ archaide-backend/web/lobby.html
@@ -1,0 +1,153 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <title>Go WebSocket Lobby</title>
+    <style>
+        body { font-family: sans-serif; }
+        #log { height: 300px; width: 500px; border: 1px solid #ccc; overflow-y: scroll; margin-bottom: 10px; padding: 5px; }
+        .player { margin-bottom: 5px; }
+        .player strong { display: inline-block; min-width: 150px;}
+        button { margin-right: 5px;}
+    </style>
+</head>
+<body>
+    <h1>WebSocket Lobby</h1>
+    <div id="log"></div>
+    <div id="status">Connecting...</div>
+    <div id="players"><h2>Players</h2></div>
+    <div id="games"><h2>Select a Game</h2></div>
+    <div id="selected-game"></div>
+
+    <script>
+        const logDiv = document.getElementById('log');
+        const statusDiv = document.getElementById('status');
+        const playersDiv = document.getElementById('players');
+        const gamesDiv = document.getElementById('games');
+        const selectedGameDiv = document.getElementById('selected-game');
+        let ws;
+        let myClientId = '';
+        let availableGames = [];
+
+        function logMessage(message) {
+            const p = document.createElement('p');
+            p.textContent = message;
+            logDiv.appendChild(p);
+            logDiv.scrollTop = logDiv.scrollHeight; // Auto-scroll
+        }
+
+        function connect() {
+            const proto = window.location.protocol === 'https:' ? 'wss:' : 'ws:';
+            const host = window.location.host; // Nimmt Host und Port von der aktuellen Seite
+            ws = new WebSocket(`${proto}//${host}/ws`);
+
+            ws.onopen = () => {
+                statusDiv.textContent = 'Connected';
+                logMessage('WebSocket connection opened.');
+            };
+
+            ws.onmessage = (event) => {
+                logMessage(`<- Received: ${event.data}`);
+                try {
+                    const message = JSON.parse(event.data);
+                    handleMessage(message);
+                } catch (e) {
+                    logMessage(`Error parsing message: ${e}`);
+                }
+            };
+
+            ws.onclose = () => {
+                statusDiv.textContent = 'Disconnected. Attempting to reconnect...';
+                logMessage('WebSocket connection closed. Reconnecting in 3 seconds...');
+                ws = null;
+                setTimeout(connect, 3000); // Versuche erneut zu verbinden
+            };
+
+            ws.onerror = (error) => {
+                statusDiv.textContent = 'Connection Error';
+                logMessage(`WebSocket error: ${error.message || 'Unknown error'}`);
+                // onclose wird normalerweise auch ausgelöst
+            };
+        }
+
+        function handleMessage(message) {
+            switch (message.type) {
+                case 'welcome':
+                    myClientId = message.payload.clientId;
+                    availableGames = message.payload.currentGames || [];
+                    statusDiv.textContent = `Connected as ${myClientId}`;
+                    updateGameSelectionUI();
+                    break;
+                case 'update_lobby':
+                    updatePlayerList(message.payload.players);
+                    break;
+                case 'game_selected':
+                    selectedGameDiv.textContent = `Game selected by server: ${message.payload.selectedGame}! Waiting for game start...`;
+                    // Hier könnte man die Spielauswahl-Buttons deaktivieren etc.
+                    // Reset game buttons after a delay?
+                    setTimeout(() => {
+                         selectedGameDiv.textContent = '';
+                         updateGameSelectionUI(); // Buttons wieder aktivieren
+                     }, 5000); // Nach 5 Sek wieder Auswahl ermöglichen (nur Beispiel)
+                    break;
+                case 'error':
+                     logMessage(`Server Error: ${message.payload.message}`);
+                     alert(`Server Error: ${message.payload.message}`);
+                     break;
+                default:
+                    logMessage(`Unknown message type: ${message.type}`);
+            }
+        }
+
+         function updatePlayerList(players) {
+            playersDiv.innerHTML = '<h2>Players</h2>'; // Clear old list
+            if (players) {
+                for (const [clientId, score] of Object.entries(players)) {
+                    const playerDiv = document.createElement('div');
+                    playerDiv.className = 'player';
+                    let name = clientId;
+                    if (clientId === myClientId) {
+                        name += ' (You)';
+                    }
+                    playerDiv.innerHTML = `<strong>${name}:</strong> ${score} points`;
+                    playersDiv.appendChild(playerDiv);
+                }
+            }
+        }
+
+        function updateGameSelectionUI() {
+            gamesDiv.innerHTML = '<h2>Select a Game</h2>'; // Clear old buttons
+            if (availableGames.length > 0) {
+                availableGames.forEach(game => {
+                    const button = document.createElement('button');
+                    button.textContent = game;
+                    button.onclick = () => selectGame(game);
+                    gamesDiv.appendChild(button);
+                });
+            } else {
+                gamesDiv.innerHTML += '<p>No games available currently.</p>';
+            }
+        }
+
+        function selectGame(gameName) {
+            if (!ws || ws.readyState !== WebSocket.OPEN) {
+                logMessage('Not connected.');
+                return;
+            }
+            const payload = { game: gameName };
+            const message = {
+                type: 'select_game',
+                payload: payload
+            };
+            const messageString = JSON.stringify(message);
+            logMessage(`-> Sending: ${messageString}`);
+            ws.send(messageString);
+            selectedGameDiv.textContent = `You selected: ${gameName}. Waiting for others...`;
+            // Deaktiviere Buttons nach Auswahl
+            gamesDiv.querySelectorAll('button').forEach(btn => btn.disabled = true);
+        }
+
+        // Initial connection
+        connect();
+    </script>
+</body>
+</html>
