--- backend/go.mod
@@ -1,8 +1,0 @@
-module github.com/Driemtax/Archaide
-
-go 1.24.2
-
-require (
-	github.com/google/uuid v1.6.0
-	github.com/gorilla/websocket v1.5.3
-)
